jira_copy_summary: look up ticket key in page header
The way JIRA navigation works in browsers, <meta> tag in <head> is not
always updated with the latest ticket, opened in a tab.  On the other
hand user-readable HTML <body> part is, for obvious reasons, properly
updated.  "Copy Summary" button uses a <meta> tag for the ticket key and
tag '#summary-val' for summary of the ticket.[1] This leads to "Copy
Summary" button putting an inconsistent text into the clipboard.

So use data attribute 'issue-key' instead, which is present in one of
the children of the tag '#dx-issuekey-val-h1'.  It is a sibling of
'#summary-val', so it gets updated with the HTML body.

[1] In current layout '#summary-val' is inside '.aui-page-header-main'.