pr-commit-links: support personal repositories
Bitbucket has two kinds of repositories, which have slightly different
URL structure.  For repositories that are part of a project:

    /projects/<project slug>/repos/<repository slug>/...

For personal repositories:

    /users/<user login>/repos/<repository slug>/...

Accomodate personal repositories by parsing both "/projects/" and
"/users" kinds of URLs.  Capture which kind was parsed and use it in the
URL of the AJAX request.