README: mark scripts & styles for support of cloud
In recent commits, support for cloud-based versions of Bitbucket [1] and
Jira was added to some of the scripts and styles [2].

Add a disclaimer to README.md that most scripts only support the
self-hosted variant.  Mark the scripts and styles that do support the
cloud-based variant as such.  Rephrase the now obsolete explicit mention
of Bitbucket Server and Jira Server in the instructions for setting
custom domain matching.

[1] I'm not sure that bitbucket.org _is_ Bitbucket Cloud, but seems so
    from how trying to get a free Bitbucket Clound instance at
    https://www.atlassian.com/software/bitbucket/bundle redirects to
    https://bitbucket.org
[2] 6b39862 ("bitbucket-pr-avatars: support bitbucker.org", 2023-02-01)
    and 91e3344 ("jira-project-avatar-as-favicon: support Jira
    Software", 2023-02-01)