confluence-simple-floating-toc: support Confluence Cloud
The style almost works with Confluence Cloud, except that the layout is
ever so slightly different that it causes the floating table of contents
to be hidden under the header of the page, even with top margin of 6em.
Add a user-controlled top margin with default to 12em, which would be
better for Cloud version and just fine for Server version.

Side note: this brings total number of user scripts and user styles in
Atlassian Tweaks that support Cloud versions to 6 out of 15.