bitbucket-fix-breadcrumbs-selection: support Bitbucket Cloud
HTML layout in Bitbucket Cloud makes it a bit harder to accidentally
select the pill-shaped (lozenge-shaped, if you prefer) markers near the
file paths, but it is still possible.

Add selector for CSS classes of lozenges that are used by Bitbucket
Cloud to user style bitbucket-fix-breadcrumbs-selection.user.css.

I have tested this change on:

  - https://bitbucket.org/andreyrybak/atlassian-tweaks/commits/b2c3b6232efc1ab6576e2fc7f5ec923a26e18283
  - https://bitbucket.org/andreyrybak/atlassian-tweaks/commits/aa0400ebd203b3e63c1ce9b3cc75111ddcc36dab

Number of scripts and styles that support Cloud variants: 7/15.