*: fix @namespace
I've found out that not changing the default value of `@namespace` might
not be ideal.[1][2]  Change it to a proper value.

Changing `@namespace` like this could be problematic -- users might
recieve it as a separate script on an update, but it shouldn't be a big
problem, they can just delete the old version.  Changing the
`@namespace` doesn't seem to affect the stored values both for scripts
and for styles.

[1] https://github.com/openstyles/stylus/wiki/Writing-UserCSS#namespace
[2] https://stackoverflow.com/a/386616/1083697