bitbucket-fix-breadcrumbs-selection: more classes for Bitbucket Server
Make selectors for Bitbucket Server more aggressive -- make the parent
node of the lozenges also non-selectable.

Unfortunately, this still isn't enough to convince Google Chrome to not
include "MODIFIED" in the copied rich content.  Plain text is fine.

Firefox doesn't seem to have this problem, but Firefox also doesn't
preserve the boldness of the filename, which is present in Bitbucket
Server.  It seems that Firefox just doesn't support rich text copying in
this way.