bitbucket-copy-commit-reference: add missing call to removeExistingContainer()
Even though this doesn't seem to affect the functionality of the
userscript on Bitbucket Cloud with its fancy on-the-fly page loading,
make sure to not leave bits of the userscript behind, just in case page
loading of Bitbucket Cloud becomes more granular.  Function
removeExistingContainer() already exists, I just forgot to call it from
function ensureLink().