bitbucket-readable-commit-messages: avoid commit lists
Style bitbucket-readable-commit-messages.user.css uses
regexp("https://bitbucket.org/.*/commits/.+") to avoid being enabled on
the page ".../commits/" of repositories.[1]  However, this isn't enough,
because there are listings of commits with URLs that don't end with
"/commits/", like the next page of the commits list.[2]

Exclude commit list pages in the selectors for commit message container
by using attribute "data-qa" which is only present on the pages for
lists of commits.  There isn't much going on in HTML of commit pages, so
we have to use negation with a ":not(...)" selector.

[1] e.g. https://bitbucket.org/andreyrybak/atlassian-tweaks/commits/
[2] e.g. https://bitbucket.org/andreyrybak/atlassian-tweaks/commits/?page=2