*.user.js: add `@license` to metadata
Only three userscripts of Atlassian Tweaks are missing the `@license`
entry in the metadata block.[1]  All userscripts also carry the whole
text of the MIT License in the source code, so there's no problem in the
sense that the source code was always licensed in an obvious
human-readable manner.  However, missing `@license` in metadata causes
Greasy Fork to show the license as "N/A" in its UI.

Add the `@license` entry to metadata blocks of these userscripts to
allow Greasy Fork [2] to parse the license from metadata.

Userscript bitbucket-copy-commit-reference.user.js isn't relevant at the
moment (but will be again in the near future), so skip incrementing its
version.

[1] Found using the following Bash incantation:
	diff <(git grep -l '@name ') <(git grep -l '@license ')
[2] Only the Confluence userscript is published there at the moment.
    https://greasyfork.org/en/scripts/470384-confluence-copy-link-buttons