bitbucket-copy-commit-reference: use ccr-lib
Userscript bitbucket-copy-commit-reference.user.js in Atlassian Tweaks
was the first one in a series of userscripts for "Copy commit reference"
button on Git hosting websites.  It was, of course, specific to
Bitbucket Server and Bitbucket Cloud.

The code from this userscript and a second userscript for GitHub was
adapted into one mega-userscript copy-commit-reference.user.js.  After a
while, the mega-userscript gained support for eight different hosting.
This mega-userscript was deemed to not be a very good idea for
various reasons.[1]

Replace the previously obsoleted [2] implementation of userscript
bitbucket-copy-commit-reference.user.js with code, adapted from the
mega-script to use the library copy-commit-reference-lib.js from [3].

Bulk of the code of classes GitHosting, BitbucketCloud, and
BitbucketServer was developed as part of the mega-script.[4]  Look for
mentions of Bitbucket in the commit messages in the log of that file for
details.[5]

Revert changes to README.md from commit [2].  Since I want to distribute
the new incarnation of copy-commit-reference userscripts under license
AGPL-3.0-only, add its text to the repository and rewrite the paragraph
in README.md about the licenses, since we now have two of them.

[1] https://github.com/rybak/copy-commit-reference-userscript/issues/14
[2] e809afe (bitbucket-copy-commit-reference: mark as obsolete,
    2023-08-28)
[3] https://github.com/rybak/copy-commit-reference-userscript
[4] The mega-script and the lib have similar filenames, don't confuse
    them.
[5] https://github.com/rybak/copy-commit-reference-userscript/commits/main/copy-commit-reference.user.js