jira_copy_summary: support Jira 9
HTML layout of Jira 9 is compatible with the code of userscript
jira_copy_summary.user.js for Jira 8.  Add version "9" as a fall-through
case together with "8" in the `switch` in function createButton.
Function createButtonForJira8 is now used both for Jira 8 and Jira 9, so
rename it to createButtonForJira8and9.

Tested on bugs.mojang.com, which uses Jira version 9.9.1:

  - https://bugs.mojang.com/browse/MC-179867
  - https://bugs.mojang.com/browse/MC-245916

Close https://github.com/rybak/atlassian-tweaks/pull/14
Helped-by: Andrei Rybak <rybak.a.v@gmail.com>