*.user.js: add `@homepageURL` to metadata
Two userscripts: bitbucket-copy-commit-reference and
confluence-space-avatar-favicons are missing `@homepageURL`.  The
userscript jira_copy_summary is using `@homepage` in its metadata.
This issues were found using the new script:

	$ ./maintenance/check-metadata.sh

It seems that extension Violentmonkey doesn't support `@homepage`, but
Tampermonkey does.[1]

Use `@homepageURL` with a link to the GitHub repository in all
userscripts, both for consistency and for better compatibility.

[1] Compare documentation at
    - https://violentmonkey.github.io/api/metadata-block/#homepageurl
    and
    - https://www.tampermonkey.net/documentation.php?locale=en#meta:homepage