generate_badges: new maintenance script
This is the script that I've used to produce the Markdown references in
the previous commits for adding Greasy Fork badges to the README.  This
Bash script uses file "greasy-fork-urls.txt" as input.  The input file
was produced manually from output of the following JavaScript snippet on
the Greasy Fork page for the set [1]:

	document
		.querySelectorAll('#browse-script-list li')
		.forEach(li =>
			console.log(li.querySelector('.script-link').href)
		)

In the future, if needed, the script can also be updated to produce
references for UserStyles.world badges, which differ from Greasy Fork
badges in color (#670000 vs #193652, see variable base_url).

[1] https://greasyfork.org/en/scripts?language=all&set=585074