bitbucket-green-merge-button: better green for Server
For consistency, replace default "green" with the color from CSS
variable --aui-lozenge-success-bg-color, which can be seen on some of
the pages of Bitbucket Server.

This is basically simpler replacement of `green` with the color
`#00875a` on top of a revert of the previous commit.  Diff between v1
and v3 of the userstyle is only in the color.  I didn't do it in v2,
because I didn't have access to a Bitbucket Server instance at the time
to find an appropriate CSS color.