README: link to sections at the top
Instead of linking to Atlassian's product pages at the top, link to the
sections of the README for "Bitbucket", "Confluence", and "Jira".  This
will provide at least some navigation to users who don't know about
GitHub's table of contents.  This will also help on bitbucket.org, which
doesn't have integrated table of contents for the README.md file.

Bitbucket mangles section names, and builtin IDs are prefixed with
`markdown-header-`.  Use hosting-provider-agnostic manually inserted
anchors with IDs instead for consistent linking to the sections.