bitbucket-copy-commit-reference: update CSS for Cloud
After an update of Bitbucket Cloud, the mangled CSS class names, which
are used by bitbucket-copy-commit-reference.user.js have changed.
Update the CSS classes used for the `button` by method wrapButton in
class BitbucketCloud.  The CSS classes for the `icon` in the same method
haven't been affected, as far as I can tell.