jira-pr-links-improver: support Jira 9
Jira 9 has the same layout as Jira 8, so similarly to recently added
support of Jira 9 to jira_copy_summary.user.js [1], we can just use the
same code as for Jira 8.

The same `switch` statement for different versions of Jira also has a
`default:` case.  This default case assigns the value for variable
`header`, but doesn't use it and `return`s instead.  This seems as an
oversight.  I probably meant to use a `break` here to let the script
finish working even on quote-unquote "unsupported" version of Jira.
Usage of `warn` instead of `error` for the log message also supports
this hypothesis.  Let the userscript work on all versions of Jira, even
those that aren't considered "supported".

[1] 962cc3a (jira_copy_summary: support Jira 9, 2023-10-03)