check-metadata.sh: ensure uniquness of key-value pairs
Script "maintenance/check-metadata.sh" checks the presence of various
key-value pairs in metadata of userscripts [1] and userstyles [2].
However, it doesn't check the number of occurrences, which makes it
possible to have two `@license`s for example.

Upgrade the script check-metadata.sh to ensure that every key-value
pair/tag in metadata of userscripts and userstyles is present exactly
once.  Because functions has_namespace and has_homepageURL also check
correctness of the value, add extra checks without the value, so that
the script "check-metadata.sh" would complain about same key with
different values.

Closes https://github.com/rybak/atlassian-tweaks/issues/16

[1] https://violentmonkey.github.io/api/metadata-block/
[2] https://github.com/openstyles/stylus/wiki/Writing-UserCSS#metadata