confluence-space-avatar-favicons: better loading listener
The listener for 'load' event seems to not be doing its job for some
reason, and it might be better to avoid it.  Quote from Stack Overflow:

	"load"

	A very different event, **load**, should only be used to detect
	a *fully-loaded page*. It is an incredibly popular mistake to
	use load where DOMContentLoaded would be much more appropriate,
	so be cautious.

<https://stackoverflow.com/a/36096571/1083697>

Rewrite the loading listener of confluence-space-avatar-favicons.user.js
to use a combination of checking document.readyState and a listener for
DOMContentLoaded event, as per recommendation from Stack Overflow.  This
combination is a recommended replacement for jQuery's $(document).ready
from <https://youmightnotneedjquery.com>.