bitbucket-readable-commit-messages: support commits not on a branch
Commits not on a branch have an additional <div> added to them with the
message "This commit is not reachable from any branch or tag in this
repository. It may be from a fork outside of this repository." The extra
<div> breaks the selector which uses nth-child pseudo-class in userstyle
bitbucket-readable-commit-messages.user.css.

I noticed that the <div> with commit message doesn't have a `class`
attribute at all.  Rewrite the selector not to rely on the order of
elements.

Examples for tests:

  - Commit on a branch:
    https://bitbucket.org/andreyrybak/atlassian-tweaks/commits/d1ff7b1d2f4f9d6fed7c706f330eaa11c73d264f

  - Commit not on a branch:
    https://bitbucket.org/andreyrybak/atlassian-tweaks/commits/19ca4f537e454e15f4e3bf1f88ebc43c0e9c559a