bitbucket-copy-commit-reference: BitbucketServer: escape HTML in convertPlainSubjectToHtml
A bug in the copy-commit-reference-lib.js caused incorrect HTML to be
generated when Jira issues or pull requests are linked in the commit
message.

Migrate to the new version of the library with the fix.  Update HTML
handling in the method convertPlainSubjectToHtml of class
BitbucketServer according to the new documentation of the class
GitHosting in the library code.