jira-pr-links-improver: show project/repo slugs
There are projects, which commit to several different repositories under
the same Jira issue.  To make distinguishing pull request links added by
the userscript jira-pr-links-improver.user.js easier, show the so called
"slugs" of the Bitbucket project and of the repository before the links
in the list items in the panel.

Rewrite creation of the links from jQuery to use plain JavaScript, while
we're here.

Reported-by: Dmitry Trubin <1437843+dtruebin@users.noreply.github.com>