bitbucket-pr-avatars: support Bitbucket Server 8.19.* better
There seems to have been an HTML layout change in a recent Bitbucket
Server version (somewhere between 8.9.* and 8.19.* [1]), which broke the
userscript bitbucket-pull-request-avatars.user.js.

In cases when userAvatar.style.backgroundImage doesn't contain a URL
that we need for the favicon, check the HTML attribute userAvatar.src,
which may exist in cases when userAvatar is an `<img>` tag.

[1] This is a very approximate range based on the fact that we needed
    changes in other scripts and styles for Bitbucket Server 8.9.*:
    - 5ca43d7 (bitbucket-copy-commit-reference: fix Jira integration for
      Bitbucket Server 8.9.*, 2024-09-06)
    - 7f3790b (bitbucket-copy-commit-reference: support Bitbucket Server
      8.9.*, 2024-09-06)
    - 11f7f2b (pr-commit-links: support Bitbucket Server v8.9.*,
      2021-11-05)

    and for Bitbucket Server 8.19.*
    - 0f6d417 (bitbucket-fast-and-furious: update selectors for
      Bitbucket Data Center, 2025-05-16)