jira-inline-code: fix `line-height` for newer Jira versions
Newer versions of Jira Software have upgraded the major number in the
attribute `data-version` of the <body> tag.  Fix the `line-height` for
Jira with this version as well, as was done previously in commit 36eebc5
("jira-inline-code: improve readability for JIRA 8", 2022-10-24).

Replace the percentage value (discouraged) with the preferred unitless
number.  See MDN docs for details:

	https://developer.mozilla.org/en-US/docs/Web/CSS/Reference/Properties/line-height